# Optimize ice.Agent.getBestPair()