@Sean-Der 2019-03-22
Hey @kixelated 

If you add a deadline to the tests what does the deadlock look like? If it is deep I am totally good with merging, but would be nice to not have two ways to do one thing. I have some bandwidth to help also.