@kixelated 2019-03-22
@Sean-Der I wanted to fix it to but it really sucks that the unit tests take a minute to run. In this case, I would change each `run` to `runInner` and see if the tests still passed. The code makes some massive assumptions about which goroutines run which functions (and if they need a lock).

I'll post the stack traces below.